Removed API deprecated since many releases
