+++ xcube/core/dsio.py
@@ -46,6 +46,8 @@ _DEPRECATION_REASON = "Functionality is redundant. Use xcube.core.store API inst
 _DEPRECATION_VERSION = "0.12.1"
 
 
+# Note, we cannot remove deprecated code as long as
+# xcube.core.xarray.DatasetAccessor.open() is using it.
 @deprecated(_DEPRECATION_REASON, version=_DEPRECATION_VERSION)
 def open_cube(input_path: str, format_name: str = None, **kwargs) -> xr.Dataset:
     """Open a xcube dataset from *input_path*.
